P10_03_chatbot/webapp/requirements.txt
botbuilder-core
botbuilder-applicationinsights
botbuilder-integration-aiohttp
botbuilder-integration-applicationinsights-aiohttp
datatypes-date-time
botbuilder-dialogs
botbuilder-ai
botbuilder-testing
aiohttp
python-dotenv
